Refactor implementation of Scheme derived expressions
The method to register and to detect Scheme derived expressions has
become simplier, more flexible, and more easily extensible.  Similiar
to the scheme library mechanism, the set of derived expressions are
now registered with and found using a "syntactic" environment during
Scheme code analysis.  This refactoring of the implementation is a
first step towards implementing a Scheme macro system.

The CSCM roadmap has been modified to raise the priority of the Scheme
macro system and lower the priority of the Scheme numeric tower.